+++ components/04-navigation/_nav.scss
@@ -70,7 +70,7 @@
     margin: 0;
     padding: 0;
     @include mq(xs,s) {
-      padding: .5rem 0;
+      padding: 0 0 .5rem 0;
     }
   }
   .nav__item {
@@ -81,6 +81,9 @@
     text-align: center;
     &:first-child {
       margin-left: 0;
+      @include mq(xs,m) {
+        padding-top: 0;
+      }
     }
     @include mq(m) {
       margin: 0 0 0 1rem;
