+++ components/04-navigation/main-nav.js
@@ -4,8 +4,10 @@ import Nav from '../../assets/js/nav'
 domready(() => {
     const btn = document.querySelector('.js-nav-btn')
     const nav = document.querySelector('.js-main-nav')
+
+    if (!btn || !nav) return false
+
     const mainNav = new Nav(btn,nav)
-    if (btn) {
-        btn.addEventListener('click', e => mainNav.toggleNav(e))
-    }
+    
+    btn.addEventListener('click', e => mainNav.toggleNav(e))
 })
