+++ components/header/header.hbs
@@ -2,7 +2,7 @@
     <div class="header__top" role="banner">
         <div class="container">
             <div class="grid grid--gutterless">
-                <div class="grid__col col-7@xs col-5@m">
+                <div class="grid__col col-5@m">
                     <div class="logo">
                         <a href="#1"><img src="../../assets/img/ons-logo-white.svg" alt="Office for National Statistics logo" class="logo__img header__logo"></a>
                     </div>
